requirements_local.txt
# Dependencias para entorno local (Jetson Nano)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyserial==3.5
pymodbus==3.5.2
supabase==2.0.2
psycopg2-binary==2.9.7
requests==2.31.0
python-multipart==0.0.6
pydantic==2.4.2
asyncio-mqtt==0.16.1
netifaces==0.11.0
scapy==2.5.0
python-dotenv==1.0.0
